requirements.txt
streamlit>=1.30.0
openai>=1.10.0
anthropic>=0.18.0
google-generativeai>=0.3.2
pandas>=2.0.0
plotly>=5.18.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0